Rename replay_skipped_bets to replay_pending_bets
